[ci skip] refactor: Change folder name (Tensor -> Tensors)
