+++ Sources/cubbydnn/Tensors/TensorData.cpp
@@ -4,7 +4,7 @@
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
-#include <cubbydnn/Tensor/TensorData.hpp>
+#include <cubbydnn/Tensors/TensorData.hpp>
 
 #include <utility>
