+++ Sources/cubbydnn/Tensors/TensorInfo.cpp
@@ -4,7 +4,7 @@
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
-#include <cubbydnn/Tensor/TensorInfo.hpp>
+#include <cubbydnn/Tensors/TensorInfo.hpp>
 
 #include <utility>
