+++ Sources/cubbydnn/Tensors/TensorObject.cpp
@@ -4,7 +4,7 @@
 // personal capacity and are not conveying any rights to any intellectual
 // property of any third parties.
 
-#include <cubbydnn/Tensor/TensorObject.hpp>
+#include <cubbydnn/Tensors/TensorObject.hpp>
 
 #include <cassert>
